separating master steps and removing un used classes
