+++ cg_lims/app/schemas/sample.py
@@ -2,8 +2,6 @@ from pydantic import BaseModel
 import datetime as dt
 from typing import Optional
 
-from cg_lims.app.schemas.workflows.twist_3 import Twist3
-
 
 class Sample(BaseModel):
     id: str
@@ -15,4 +13,3 @@ class Sample(BaseModel):
     prepared_date: Optional[dt.date]
     sequenced_date: Optional[dt.date]
     delivery_date: Optional[dt.date]
-    twist: Optional[Twist3]
